Refactor 'learning paths' to 'learning center'
