+++ frontend/src/app/Routes.tsx
@@ -7,7 +7,7 @@ const InstalledApplications = React.lazy(
 const ExploreApplications = React.lazy(
   () => import('../pages/exploreApplication/ExploreApplications'),
 );
-const LearningPathsPage = React.lazy(() => import('../pages/learningPaths/LearningPaths'));
+const LearningCenterPage = React.lazy(() => import('../pages/learningCenter/LearningCenter'));
 const NotFound = React.lazy(() => import('../pages/NotFound'));
 
 const Routes: React.FC = () => (
@@ -17,7 +17,7 @@ const Routes: React.FC = () => (
     <Switch>
       <Route path="/" exact component={InstalledApplications} />
       <Route path="/explore" exact component={ExploreApplications} />
-      <Route path="/learningpaths" exact component={LearningPathsPage} />
+      <Route path="/learning-center" exact component={LearningCenterPage} />
       <Route component={NotFound} />
     </Switch>
   </React.Suspense>
