+++ frontend/src/pages/learningCenter/LearningCenterFilter.scss
