+++ frontend/src/pages/learningCenter/learningCenterUtils.ts
